Replace int with time_t in CTime to fix Y2038 overflow
_time_val was stored as int (32-bit), causing overflow for dates after
Jan 19, 2038. Replace all time-related int types with time_t (64-bit)
throughout CTime and its callers. Remove static_cast<int> wrappers
around time()/mktime() calls, and eliminate intermediate int casts in
the static helpers (EibTime, EibGMTime, etc.). Also rename secTo() to
SecondsTo() for clarity, and fix CString::operator+= ambiguity in
HttpSession.cpp by casting to int64.

+++ EIBStdLib/src/HttpSession.cpp
@@ -28,7 +28,7 @@ void CSessionManager::GenerateNewSession(const CString& user_name,CString& sessi
 	tmp += "_";
 	CTime now;
 	now.SetNow();
-	tmp += now.GetTime();
+	tmp += (int64)now.GetTime();
 
 	CDigest digest;
 	digest.HashString(tmp, session_id);
